MB-6237 Rename mixins to hosts to better reflect the content
